Removed redundant custom FXML instantiation.
I'm not sure how long ago this applied as I can't remember compiling
successfully without instantiating custom controls within the controller
itself as of JDK8_u20. Maybe the last update provided this
functionality? Whatever works. It's more intuitive this way as the FXML
is supposed to be instantiated before the controller is attached.